% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GcdNum.R
\name{hgch_choropleth_Gcd}
\alias{hgch_choropleth_Gcd}
\title{choropleth chart Gcd}
\usage{
hgch_choropleth_Gcd(data = NULL, ...)
}
\description{
\code{hgch_choropleth_Gcd()} Create a Highcharter choropleth map based on a particular data type.
In this case, you can load data with only one column, where it's a \strong{geocode column},
or make sure that the first column of your data meet this condition
}
\section{Ftype}{

Gcd
}

\examples{
data <- sample_data("Gcd", n = 30)
hgch_choropleth_Gcd(data)


# if you want to calculate the average instead of the sum, you can use agg inside a function
hgch_choropleth_Gcd(data,
                    agg = "mean")


}
\concept{Gcd plots}
