% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmNum.R
\name{hgch_choropleth_Gnm}
\alias{hgch_choropleth_Gnm}
\title{choropleth chart Gnm}
\usage{
hgch_choropleth_Gnm(data = NULL, ...)
}
\description{
\code{hgch_choropleth_Gnm()} Create a Highcharter choropleth map based on a particular data type.
In this case, you can load data with only one column, where it's a \strong{geoname column},
or make sure that the first column of your data meet this condition
}
\section{Ftype}{

Gnm
}

\examples{
data <- sample_data("Gnm", n = 30)
hgch_choropleth_Gnm(data)


# if you want to calculate the average instead of the sum, you can use agg inside a function
hgch_choropleth_Gnm(data,
                    agg = "mean")


}
\concept{Gnm plots}
