% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth_GnmNum.R
\name{hgch_choropleth_GnmNum}
\alias{hgch_choropleth_GnmNum}
\title{choropleth chart Gnm Num}
\usage{
hgch_choropleth_GnmNum(data = NULL, ...)
}
\description{
\code{hgch_choropleth_GnmNum()} Create a Highcharter choropleth map based on a particular data type.
In this case, you can load data with only two columns, where the firts it's a \strong{geoname column},
and the second is a \strong{numeric class column}, or make sure that the first two columns of
your data meet this condition
}
\section{Ftype}{

Gnm-Num
}

\examples{
data <- sample_data("Gnm-Num", n = 30)
hgch_choropleth_GnmNum(data)


# if you want to calculate the average instead of the sum, you can use agg inside a function
hgch_choropleth_GnmNum(data,
                       agg = "mean")


}
\concept{Gnm-Num plots}
