% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hgch_theme}
\alias{hgch_theme}
\title{Default theme}
\usage{
hgch_theme(opts = NULL)
}
\arguments{
\item{opts}{A list with parameters to modify the theme.}
}
\description{
Default theme
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_alone())

}
