% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgch-prep.R
\docType{package}
\name{hgchmaps-package}
\alias{hgchmaps}
\alias{hgchmaps-package}
\title{hgchmaps: A highcharter map wrapper}
\description{
A wrapper for highcharter
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/datasketch/hgchmaps/issues}
}

}
\keyword{internal}
